Remove redundant classloader wrapper from exclude_columns
Engine-provided functions do not need
`ClassLoaderSafeConnectorTableFunction` wrapper.
