+++ core/trino-main/src/main/java/io/trino/connector/system/SystemConnectorModule.java
@@ -30,7 +30,7 @@ import io.trino.connector.system.jdbc.TableJdbcTable;
 import io.trino.connector.system.jdbc.TableTypeJdbcTable;
 import io.trino.connector.system.jdbc.TypesJdbcTable;
 import io.trino.connector.system.jdbc.UdtJdbcTable;
-import io.trino.operator.table.ExcludeColumns;
+import io.trino.operator.table.ExcludeColumnsFunction;
 import io.trino.operator.table.Sequence;
 import io.trino.spi.connector.SystemTable;
 import io.trino.spi.function.table.ConnectorTableFunction;
@@ -79,7 +79,7 @@ public class SystemConnectorModule
         binder.bind(GlobalSystemConnector.class).in(Scopes.SINGLETON);
 
         Multibinder<ConnectorTableFunction> tableFunctions = Multibinder.newSetBinder(binder, ConnectorTableFunction.class);
-        tableFunctions.addBinding().toProvider(ExcludeColumns.class).in(Scopes.SINGLETON);
+        tableFunctions.addBinding().to(ExcludeColumnsFunction.class).in(Scopes.SINGLETON);
         tableFunctions.addBinding().toProvider(Sequence.class).in(Scopes.SINGLETON);
     }
 }
