+++ core/trino-main/src/main/java/io/trino/metadata/GlobalFunctionCatalog.java
@@ -21,7 +21,7 @@ import com.google.errorprone.annotations.ThreadSafe;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import io.trino.connector.system.GlobalSystemConnector;
-import io.trino.operator.table.ExcludeColumns.ExcludeColumnsFunctionHandle;
+import io.trino.operator.table.ExcludeColumnsFunction.ExcludeColumnsFunctionHandle;
 import io.trino.operator.table.Sequence.SequenceFunctionHandle;
 import io.trino.operator.table.json.JsonTable.JsonTableFunctionHandle;
 import io.trino.spi.function.AggregationFunctionMetadata;
@@ -55,7 +55,7 @@ import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static io.trino.metadata.OperatorNameUtil.isOperatorName;
 import static io.trino.metadata.OperatorNameUtil.mangleOperatorName;
 import static io.trino.metadata.OperatorNameUtil.unmangleOperator;
-import static io.trino.operator.table.ExcludeColumns.getExcludeColumnsFunctionProcessorProvider;
+import static io.trino.operator.table.ExcludeColumnsFunction.getExcludeColumnsFunctionProcessorProvider;
 import static io.trino.operator.table.Sequence.getSequenceFunctionProcessorProvider;
 import static io.trino.operator.table.json.JsonTable.getJsonTableFunctionProcessorProvider;
 import static io.trino.spi.function.FunctionKind.AGGREGATE;
